techmart_accounting/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==25.3.0
whitenoise==6.12.0
dj-database-url==3.1.2
Pillow==10.3.0
anthropic==0.95.0
